studyperiod encoded outside of study group: https://github.com/metanorma/metanorma-iec/issues/191
